Make the appbar less dumb Originally, each layout copied the appbar code.  I've moved the appbar to a centralized location, with its settings being able to be controlled through an atom.
