+++ front-end/src/components/loading/index.ts
@@ -1,0 +1,3 @@
+import { PageLoader } from "./page-loader";
+
+export { PageLoader };
